+++ src/components/money-input/money-input.test.jsx
@@ -1,9 +1,5 @@
 import React, { createRef } from 'react';
-import {
-  fireEvent,
-  render,
-  screen,
-} from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import CustomFieldInputText from '../custom-field-input-text/custom-field-input-text.jsx';
 import MoneyInput from './money-input.jsx';
@@ -37,7 +33,7 @@ describe('MoneyInput', () => {
 
     // NOTE: This space is character code 160, non-breaking space. It did break my brain though
     expect(screen.getByLabelText('currency')).toHaveValue('IQD 10.111');
-    fireEvent.focus(screen.getByLabelText('currency'));
+    userEvent.click(screen.getByLabelText('currency'));
     expect(screen.getByLabelText('currency')).toHaveValue(10.111);
   });
 
@@ -52,14 +48,6 @@ describe('MoneyInput', () => {
         expect(currencyProps).toContain(key);
       });
     });
-
-    it('presents contextual error state', () => {
-      const errorText = 'What do you want from us monster!?';
-      render(<MoneyInput {...requiredProps} value={350} errorText={errorText} />);
-
-      expect(screen.getByLabelText('currency')).toBeInvalid();
-      expect(screen.getByLabelText('currency')).toHaveDescription(errorText);
-    });
   });
 
   describe('dirty ref API', () => {
@@ -69,7 +57,6 @@ describe('MoneyInput', () => {
       userEvent.click(screen.getByLabelText('currency'));
       expect(ref.current.dirty).toEqual(false);
       userEvent.type(screen.getByLabelText('currency'), '12');
-      userEvent.tab();
       expect(ref.current.dirty).toEqual(true);
     });
   });
@@ -87,7 +74,6 @@ describe('MoneyInput', () => {
       const onChangeSpy = jest.fn();
       const { getByLabelText } = render(<MoneyInput {...requiredProps} onChange={onChangeSpy} />);
       userEvent.type(getByLabelText('currency'), '1');
-      userEvent.tab();
       expect(onChangeSpy.mock.calls.length).toEqual(1);
     });
   });
@@ -100,7 +86,7 @@ describe('MoneyInput', () => {
 
     it('does not enter edit mode on focus', () => {
       render(<MoneyInput {...requiredProps} readOnly={true} />);
-      fireEvent.focus(screen.getByLabelText('currency'));
+      userEvent.click(screen.getByLabelText('currency'));
       expect(screen.getByLabelText('currency')).toHaveAttribute('type', 'text');
     });
 
@@ -114,9 +100,9 @@ describe('MoneyInput', () => {
     it('does not trigger error state for valid value', () => {
       render(<MoneyInput {...requiredProps} />);
 
-      fireEvent.focus(screen.getByLabelText('currency'));
-      fireEvent.change(screen.getByLabelText('currency'), { target: { value: 1234 } });
-      fireEvent.blur(screen.getByLabelText('currency'));
+      userEvent.click(screen.getByLabelText('currency'));
+      userEvent.type(screen.getByLabelText('currency'), '1234');
+      userEvent.tab();
 
       expect(screen.getByLabelText('currency')).toBeValid();
       expect(screen.getByLabelText('currency')).toHaveValue('$1,234.00');
@@ -125,26 +111,12 @@ describe('MoneyInput', () => {
     it('does not switch to view mode when its value is numerically invalid', () => {
       render(<MoneyInput {...requiredProps} />);
 
-      fireEvent.focus(screen.getByLabelText('currency'));
-      fireEvent.change(screen.getByLabelText('currency'), { target: { value: 12.111111 } });
-      fireEvent.blur(screen.getByLabelText('currency'));
-
+      userEvent.click(screen.getByLabelText('currency'));
+      userEvent.type(screen.getByLabelText('currency'), '12.111111');
       expect(screen.getByLabelText('currency')).toHaveAttribute('type', 'number');
+
+      userEvent.tab();
       expect(screen.getByLabelText('currency')).toHaveValue(12.111111);
-    });
-
-    xit('does not switch to view mode when its value is numerically invalid', () => {
-      // This test does not work because programmatically setting value
-      // on a number input does not reflect a user typing.
-      // See commit for more details.
-      render(<MoneyInput {...requiredProps} />);
-
-      fireEvent.focus(screen.getByLabelText('currency'));
-      fireEvent.change(screen.getByLabelText('currency'), { target: { value: '12..' } });
-      fireEvent.blur(screen.getByLabelText('currency'));
-
-      expect(screen.getByLabelText('currency')).toHaveAttribute('type', 'number');
-      expect(screen.getByLabelText('currency')).toHaveValue('12..');
     });
 
     it('accepts an undefined value', () => {
@@ -200,35 +172,33 @@ describe('MoneyInput', () => {
       const ref = createRef(null);
       render(<MoneyInput {...requiredProps} id="test-input" label="Test label" ref={ref} />);
 
-      fireEvent.focus(screen.getByLabelText('Test label'));
-      fireEvent.change(screen.getByLabelText('Test label'), { target: { value: 1234 } });
+      userEvent.click(screen.getByLabelText('Test label'));
+      userEvent.type(screen.getByLabelText('Test label'), '12.34');
+      expect(ref.current.value).toStrictEqual(1234);
       userEvent.tab();
-      expect(ref.current.value).toStrictEqual([123400, 'USD']);
-      fireEvent.blur(screen.getByLabelText('Test label'));
-      expect(ref.current.value).toStrictEqual([123400, 'USD']);
+      expect(ref.current.value).toStrictEqual(1234);
     });
 
     it('can get a negative value', () => {
       const ref = createRef(null);
       render(<MoneyInput {...requiredProps} id="test-input" label="Test label" ref={ref} />);
 
-      fireEvent.focus(screen.getByLabelText('Test label'));
-      fireEvent.change(screen.getByLabelText('Test label'), { target: { value: -1234 } });
+      userEvent.click(screen.getByLabelText('Test label'));
+      userEvent.type(screen.getByLabelText('Test label'), '-12.34');
+      expect(ref.current.value).toStrictEqual(-1234);
       userEvent.tab();
-      expect(ref.current.value).toStrictEqual([-123400, 'USD']);
-      fireEvent.blur(screen.getByLabelText('Test label'));
-      expect(ref.current.value).toStrictEqual([-123400, 'USD']);
+      expect(ref.current.value).toStrictEqual(-1234);
     });
 
     it('can be undefined', () => {
       const ref = createRef(null);
       render(<MoneyInput {...requiredProps} id="test-input" label="Test label" ref={ref} />);
 
-      fireEvent.focus(screen.getByLabelText('Test label'));
-      fireEvent.change(screen.getByLabelText('Test label'), { target: { value: '' } });
+      userEvent.click(screen.getByLabelText('Test label'));
+      userEvent.type(screen.getByLabelText('Test label'), '');
 
       expect(ref.current.value).toBeUndefined();
-      fireEvent.blur(screen.getByLabelText('Test label'));
+      userEvent.tab();
       expect(ref.current.value).toBeUndefined();
     });
   });
@@ -247,6 +217,14 @@ describe('MoneyInput', () => {
       userEvent.hover(screen.getByRole('img', { name: 'More information' }));
       userEvent.unhover(screen.getByRole('img', { name: 'More information' }));
       expect(screen.getByLabelText(requiredProps.label)).toHaveDescription('');
+    });
+  });
+
+  describe('validationMessage API', () => {
+    it('shows the message', () => {
+      render(<MoneyInput {...requiredProps} validationMessage={'What do you want from us monster!?'} />);
+      expect(screen.getByLabelText('currency')).toBeInvalid();
+      expect(screen.getByLabelText('currency')).toHaveDescription('What do you want from us monster!?');
     });
   });
 });